+++ CSM Database Core/Core/Attributes/EntityRelationAttribute.cs
@@ -3,14 +3,14 @@
 namespace CSM_Database_Core.Core.Attributes;
 
 /// <summary>
-///     Attribute to mark a relation dependant.
+///     Attribute to mark a relation dependency.
 /// </summary>
 [AttributeUsage(AttributeTargets.Property, AllowMultiple = false, Inherited = true)]
-public class EntityDependantAttribute
+public class EntityRelationAttribute
     : RelationAttributeBase {
 
     /// <inheritdoc/>
-    public EntityDependantAttribute(string? name = null)
+    public EntityRelationAttribute(string? name = null)
         : base(name) {
     }
 }