+++ CSM Database Core/Entities/Abstractions/Interfaces/ICatalogEntity.cs
@@ -1,9 +1,0 @@
-﻿namespace CSM_Database_Core.Entities.Abstractions.Interfaces;
-
-
-/// <summary>
-///     Represents an <see cref="IEntity"/> with catalog identification properties.
-/// </summary>
-public interface ICatalogEntity
-    : IEntity, INamedEntity, IReferencedEntity, IActivableEntity {
-}