+++ CSM Database Core/CHANGELOG.md
@@ -1,10 +1,136 @@
 # CSM Database Core CHANGELOG
 
-## [x.x.x] - xx.xx-xxxx
+## [7.0.0] - 21.07-2026
 
 ### Changes
 
-- 
+- Now when database connection options weren't found on app assemblies a detailed console message is displayed on Design an Development time.
+
+### Breaks
+
+- Now [NamedEntity] have new properties definitions.
+  
+  - [Description]: Now it doesn't have length limit.
+  - [Name]: Now is unique along entities.
+
+- Renamed [DatabaseDesignFactoryBase] to [DatabaseDesignerBase] to keep shorter naming conventions with sense.
+- Removed [CatalogEntity] and [ReferencedEntity] since [NamedEntities] have enough properties to deteermine an entity efficient search and constant referencing.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.9           | 10.0.10         |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.9           | 10.0.10         |
+
+## [6.1.0] - 24.06-2026
+
+### Changes
+
+- Now relations are able to be added no matter what direction of relationship they have at update operations.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.6] - 18.06-2026
+
+### Fixes
+
+- Fixed [DepotBase] [Update] method issue for a safe entity relation update just when the Ids are different.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.5] - 18.06-2026
+
+### Fixes
+
+- Fixed [DepotBase] [Update] method issue for a safe entity relation update just when the Ids are different.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.4] - 18.06-2026
+
+### Fixes
+
+- Fixed [DepotBase] [Update] method issue that was causing not relation inclussion on original entity object.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.3] - 18.06-2026
+
+### Fixes
+
+- Fixed [DepotBase] [Update] method issue that was causing miss calculation of Flat relations upgrading.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.2] - 18.06-2026
+
+### Fixes
+
+- Fixed [SanitizeEntity] issue that was causing wrong relation detections for collections. Removed deprecated dependants logic.
+
+#### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.1] - 18.06-2026
+
+### Fixes
+
+- Fixed [SanitizeEntity] issue that was causing wrong relation detections for collections.
+
+#### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Foundation.Core                     | 4.0.0            | 4.0.0           |
+| Microsoft.EntityFrameworkCore           | 10.0.8           | 10.0.9          |
+| Microsoft.EntityFrameworkCore.SqlServer | 10.0.8           | 10.0.9          |
+
+## [6.0.0] - 18.06-2026
+
+### Fixes
+
+- Fixed [SanitizeEntity] issue that was causing wrong relation detections for collections.
+
+### Removed
+
+- Removed [EntityDependencyAttribute] and [EntityDependantAttribute] now we only use [EntityRelation].
 
 #### Dependencies